add projects package
