+++ mmlearn/datasets/__init__.py
@@ -5,13 +5,8 @@ from mmlearn.datasets.ego4d import Ego4DDataset
 from mmlearn.datasets.imagenet import ImageNet
 from mmlearn.datasets.librispeech import LibriSpeech
 from mmlearn.datasets.llvip import LLVIPDataset
-from mmlearn.datasets.medvqa import MedVQA
-from mmlearn.datasets.mimiciv_cxr import MIMICIVCXR
 from mmlearn.datasets.nihcxr import NIHCXR
 from mmlearn.datasets.nyuv2 import NYUv2Dataset
-from mmlearn.datasets.pmcoa import PMCOA
-from mmlearn.datasets.quilt import Quilt
-from mmlearn.datasets.roco import ROCO
 from mmlearn.datasets.sunrgbd import SUNRGBDDataset
 
 
@@ -21,12 +16,7 @@ __all__ = [
     "ImageNet",
     "LibriSpeech",
     "LLVIPDataset",
-    "MedVQA",
-    "MIMICIVCXR",
     "NIHCXR",
     "NYUv2Dataset",
-    "PMCOA",
-    "Quilt",
-    "ROCO",
     "SUNRGBDDataset",
 ]
